requirements.txt
torch
numpy
pandas
matplotlib
pillow
albumentations
tqdm
lightning
configargparse
timm
scikit-learn